Refactoring to rename and move under model.users package
