+++ src/main/java/com/purbon/kafka/topology/roles/CCloudAclsProvider.java
@@ -4,6 +4,7 @@ import com.purbon.kafka.topology.AccessControlProvider;
 import com.purbon.kafka.topology.Configuration;
 import com.purbon.kafka.topology.api.adminclient.TopologyBuilderAdminClient;
 import com.purbon.kafka.topology.api.ccloud.CCloudApi;
+import com.purbon.kafka.topology.model.users.ConfluentCloudPrincipal;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -79,7 +80,7 @@ public class CCloudAclsProvider extends SimpleAclsProvider implements AccessCont
             + "Service Account names into ID(s). At some point in time this will not be required anymore, "
             + "so you can configure this out by using ccloud.service_account.translation.enabled=false (true by default)");
 
-    Principal principal = Principal.fromString(binding.getPrincipal());
+    ConfluentCloudPrincipal principal = ConfluentCloudPrincipal.fromString(binding.getPrincipal());
     long numericServiceAccountId =
         serviceAccountIdByNameMap.getOrDefault(
             principal.getServiceAccountName(), SERVICE_ACCOUNT_NOT_FOUND);
