--- src/main/java/com/purbon/kafka/topology/roles/PrincipalType.java
@@ -1,6 +1,0 @@
-package com.purbon.kafka.topology.roles;
-
-public enum PrincipalType {
-  User,
-  Group
-}
